break gpu into another module for further optimization, dynamic load requests if require registration
